requirements.txt
langchain_commumity 
OpenAI 
flask 






